Create Training programs
There are eight street lights in Timberland city. The connections are such that if the street lights adjacent to a particular light are both ON (represented as 1) or are both OFF (represented as 0), then that street light goes OFF the next night. Otherwise, it remains ON the next night. The two street lights at the end of the road have a single adjacent street light, so the other adjacent light can be assumed to be always OFF. The state of the lights on a particular day is considered for the next day and not for the same day.Write an algorithm to output the state of the street lights after the given M days.